Tweak to use longs instead which is better for larger files

+++ Runtime/Plugins/Android/backgrounddownload.androidlib/src/main/java/com/unity3d/backgrounddownload/BackgroundDownload.java
@@ -110,8 +110,8 @@ public class BackgroundDownload {
             return 1.0f;
         }
         cursor.moveToFirst();
-        int downloaded = cursor.getInt(cursor.getColumnIndex(DownloadManager.COLUMN_BYTES_DOWNLOADED_SO_FAR));
-        int total = cursor.getInt(cursor.getColumnIndex(DownloadManager.COLUMN_TOTAL_SIZE_BYTES));
+        long downloaded = cursor.getLong(cursor.getColumnIndex(DownloadManager.COLUMN_BYTES_DOWNLOADED_SO_FAR));
+        long total = cursor.getLong(cursor.getColumnIndex(DownloadManager.COLUMN_TOTAL_SIZE_BYTES));
         cursor.close();
         if (downloaded <= 0)
             return 0.0f;
